Make more tests work
