+++ server/src/main/java/io/crate/planner/optimizer/rule/ExtractConstantJoinConditionsIntoFilter.java
@@ -6,7 +6,7 @@
  * you may not use this file except in compliance with the License.  You may
  * obtain a copy of the License at
  *
- *   http://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
@@ -22,14 +22,10 @@
 package io.crate.planner.optimizer.rule;
 
 import static io.crate.planner.optimizer.matcher.Pattern.typeOf;
-import static io.crate.planner.optimizer.matcher.Patterns.source;
-import static io.crate.planner.optimizer.rule.ExtractConstantJoinCondition.numberOfRelationsUsed;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Set;
-import java.util.function.Function;
 import java.util.function.UnaryOperator;
 
 import io.crate.analyze.relations.QuerySplitter;
@@ -38,74 +34,72 @@ import io.crate.expression.symbol.Symbol;
 import io.crate.metadata.NodeContext;
 import io.crate.metadata.RelationName;
 import io.crate.metadata.TransactionContext;
+import io.crate.planner.consumer.RelationNameCollector;
 import io.crate.planner.operators.Filter;
 import io.crate.planner.operators.JoinPlan;
+import io.crate.planner.optimizer.costs.PlanStats;
 import io.crate.planner.operators.LogicalPlan;
 import io.crate.planner.optimizer.Rule;
-import io.crate.planner.optimizer.costs.PlanStats;
-import io.crate.planner.optimizer.matcher.Capture;
 import io.crate.planner.optimizer.matcher.Captures;
 import io.crate.planner.optimizer.matcher.Pattern;
-import io.crate.sql.tree.JoinType;
 
-public class MoveFilterIntoJoin implements Rule<Filter> {
+public class ExtractConstantJoinConditionsIntoFilter implements Rule<JoinPlan> {
 
-    private final Capture<JoinPlan> joinCapture;
-    private final Pattern<Filter> pattern;
+    private final Pattern<JoinPlan> pattern;
 
-    public MoveFilterIntoJoin() {
-        this.joinCapture = new Capture<>();
-        this.pattern = typeOf(Filter.class)
-            .with(source(),
-                typeOf(JoinPlan.class)
-                    .capturedAs(joinCapture)
-            );
+    public ExtractConstantJoinConditionsIntoFilter() {
+        this.pattern = typeOf(JoinPlan.class)
+            .with(j -> j.joinCondition() != null && j.isConstandJoinConditionsExtracted() == false);
     }
 
     @Override
-    public Pattern<Filter> pattern() {
+    public Pattern<JoinPlan> pattern() {
         return pattern;
     }
 
     @Override
-    public LogicalPlan apply(Filter filter,
+    public LogicalPlan apply(JoinPlan nl,
                              Captures captures,
                              PlanStats planStats,
                              TransactionContext txnCtx,
                              NodeContext nodeCtx,
                              UnaryOperator<LogicalPlan> resolvePlan) {
-        var join = captures.get(joinCapture);
-        var query = filter.query();
-        var allConditions = QuerySplitter.split(query);
+        var conditions = nl.joinCondition();
+        var allConditions = QuerySplitter.split(conditions);
         var constantConditions = new HashMap<Set<RelationName>, Symbol>(allConditions.size());
-        var nonConstantConditions = new HashMap<Set<RelationName>, Symbol>(allConditions.size());
+        var nonConstantConditions = new HashSet<Symbol>(allConditions.size());
         for (var condition : allConditions.entrySet()) {
             if (numberOfRelationsUsed(condition.getValue()) <= 1) {
                 constantConditions.put(condition.getKey(), condition.getValue());
             } else {
-                nonConstantConditions.put(condition.getKey(), condition.getValue());
+                nonConstantConditions.add(condition.getValue());
             }
         }
-        if (nonConstantConditions.isEmpty()) {
-            return null;
-        }
-        var result = new ArrayList<Symbol>();
-        for (var entries : nonConstantConditions.entrySet()) {
-            if (new HashSet<>(join.getRelationNames()).containsAll(entries.getKey())) {
-                result.add(entries.getValue());
-            }
-        }
-        if (result.isEmpty()) {
-            return null;
+        if (constantConditions.isEmpty() || nonConstantConditions.isEmpty()) {
+            return new JoinPlan(
+                nl.lhs(),
+                nl.rhs(),
+                nl.joinType(),
+                nl.joinCondition(),
+                nl.isFiltered(),
+                nl.isRewriteFilterOnOuterJoinToInnerJoinDone(),
+                true
+            );
         } else {
-            JoinPlan newJoin = null;
-            if (join.joinType() == JoinType.CROSS) {
-                newJoin = new JoinPlan(join.lhs(), join.rhs(), JoinType.INNER, AndOperator.join(result));
-            } else {
-                result.add(join.joinCondition());
-                newJoin = new JoinPlan(join.lhs(), join.rhs(), join.joinType(), AndOperator.join(result));
-            }
-            return Filter.create(newJoin, AndOperator.join(constantConditions.values()));
+            return Filter.create(
+                new JoinPlan(
+                    nl.lhs(),
+                    nl.rhs(),
+                    nl.joinType(),
+                    AndOperator.join(nonConstantConditions),
+                    nl.isFiltered(),
+                    nl.isRewriteFilterOnOuterJoinToInnerJoinDone(),
+                    true
+                ), AndOperator.join(constantConditions.values()));
         }
+    }
+
+    static int numberOfRelationsUsed(Symbol joinCondition) {
+        return RelationNameCollector.collect(joinCondition).size();
     }
 }
